+++ CHANGELOG.md
@@ -34,12 +34,12 @@
 * tbd... `ads data models` for each detail tables update `_useless_to_keep_history` ***ONLY at `ads` parent level*** to force versioning component to generate history line and update audit fields - implementation through API code for corresponding `ads` object and operations (insert / update / delete)
 
 
-* wip... 230402piu_a '`ads data models` for each detail tables update `_useless_to_keep_history` ***ONLY at `ads` parent level*** to force versioning component to generate history line and update audit fields - implementation tried by capturing `before_flush` events at each ORM class object level. Made for:
+* wip... 230402piu_a `ads data models` for each detail tables update `_useless_to_keep_history` ***ONLY at `ads` parent level*** to force versioning component to generate history line and update audit fields - implementation tried by capturing `before_flush` events at each ORM class object level. Made for:
     * [x] decision criteria
     * [x] evaluation - NOT APPLICABLE
     * [x] general data - NOT APPLICABLE
     * [x] org_map
-    * [ ] relationships
+    * [x] relationships
     * [x] revenue - NOT APPLICABLE
     * [ ] solution